challenge5/requirements.txt
python-dotenv
semantic-kernel
azure-search-documents
openai
azure-core
azure-monitor-opentelemetry
opentelemetry-api